nemuopl refactoring
