+++ contrib/vs2015/adplug/adplug.vcxproj
@@ -77,7 +77,8 @@
     <ClCompile Include="..\..\..\src\woodyopl.cpp" />
     <ClCompile Include="..\..\..\src\xad.cpp" />
     <ClCompile Include="..\..\..\src\xsm.cpp" />
-    <ClCompile Include="..\..\..\src\opl3.c" />
+    <ClCompile Include="..\..\..\src\nemuopl.cpp" />
+    <ClCompile Include="..\..\..\src\nukedopl.c" />
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="..\..\..\src\a2m.h" />
@@ -144,7 +145,7 @@
     <ClInclude Include="..\..\..\src\xad.h" />
     <ClInclude Include="..\..\..\src\xsm.h" />
     <ClInclude Include="..\..\..\src\nemuopl.h" />
-    <ClInclude Include="..\..\..\src\opl3.h" />
+    <ClInclude Include="..\..\..\src\nukedopl.h" />
   </ItemGroup>
   <ItemGroup>
     <None Include="packages.config" />
