+++ src/nukedopl.c
